+++ src/commands/generate/rto/command.ts
@@ -1,4 +1,4 @@
-import type { ZanixFolderGenericTree } from '@zanix/types'
+import type { ZanixFolderGenericTree } from 'typings/tree.ts'
 import type { Commander } from 'cli'
 import type { FieldDef } from 'commands/generate/rto/parser.ts'