+++ src/commands/new/lib/tree/projects/server.ts
@@ -1,4 +1,4 @@
-import type { ZanixServerSrcTree } from '@zanix/types'
+import type { ZanixServerSrcTree } from 'typings/tree.ts'
 
 import { ZanixTree } from 'commands/new/lib/tree/base-tree.ts'
 import {